# Optimize the performance of loading TouchStone files.