@FranzForstmayr 2024-07-23
Maybe just this change would be enough:
```python
            line_l = line.lower()

            if line[0] in ["!", "#", "["]:
                for k, v in self._parse_dict.items():
                    if line_l.startswith(k):
                        v(line)
                        break
            else:
                ...
```