@Asachoo 2024-07-24
## Summary
Thank you, @FranzForstmayr, for your insightful suggestions. I have summarized the changes made in this PR, which includes three modifications:
1. Caching the `numbers_per_line` variable using `cached_property`;
2. Pre-checking the character type of `line_l[0]`;
3. Using `list(map(float, ...))` to replace `list comprehensions`.

## Data
Below are the performance changes for each modification applied individually:

### 1. Master Branch (BaseLine)
```bash
Time taken to load network 0.04 MB: 0.7411 ms (std: 0.0923 ms, min: 0.6660 ms)
Time taken to load network 1.37 MB: 43.9980 ms (std: 2.8768 ms, min: 42.1416 ms)
Time taken to load network 11.32 MB: 216.2328 ms (std: 11.7919 ms, min: 206.6809 ms)
Time taken to load network 78.85 MB: 1230.7935 ms (std: 47.2551 ms, min: 1182.1987 ms)
Time taken to load network 199.82 MB: 3231.9390 ms (std: 185.4683 ms, min: 3038.6736 ms)
```

### 2. Set `numbers_per_line` to be `cached_property`
```bash
Time taken to load network 0.04 MB: 0.7005 ms (std: 0.0896 ms, min: 0.6438 ms)
Time taken to load network 1.37 MB: 37.1880 ms (std: 3.5364 ms, min: 35.4734 ms)
Time taken to load network 11.32 MB: 196.8984 ms (std: 14.4792 ms, min: 181.8379 ms)
Time taken to load network 78.85 MB: 1050.9514 ms (std: 47.1150 ms, min: 1015.2888 ms)
Time taken to load network 199.82 MB: 2746.3581 ms (std: 168.1199 ms, min: 2619.1348 ms)
```

### 3. Add a preliminary checks on strings, to keep the for..else structure valid:
```python3
...
continue_go = True
if line[0] in ["!", "#", "["]:
    for k, v in self._parse_dict.items():
        if line_l.startswith(k):
            v(line)
            continue_go = False
            break
if continue_go:
    values = [float(v) for v in line.partition("!")[0].split()]
    ...
```
Output:
```bash
Time taken to load network 0.04 MB: 0.6521 ms (std: 0.0881 ms, min: 0.5911 ms)
Time taken to load network 1.37 MB: 35.4366 ms (std: 3.9613 ms, min: 33.2773 ms)
Time taken to load network 11.32 MB: 177.2632 ms (std: 14.0723 ms, min: 169.9531 ms)
Time taken to load network 78.85 MB: 1033.8643 ms (std: 57.5237 ms, min: 997.0122 ms)
Time taken to load network 199.82 MB: 2739.5001 ms (std: 177.7936 ms, min: 2595.1284 ms)
```

### 4. Use `map(float, ...)` to replace list comprehensions `[float(v) for v in ...]`
```python3
...
if "!" in line:
    line = line.partition("!")[0]
values = list(map(float, line.split()))
...
```
Output:
```bash
Time taken to load network 0.04 MB: 0.8061 ms (std: 0.1930 ms, min: 0.6724 ms)
Time taken to load network 1.37 MB: 42.8646 ms (std: 3.4950 ms, min: 40.6724 ms)
Time taken to load network 11.32 MB: 211.8912 ms (std: 13.7295 ms, min: 205.5249 ms)
Time taken to load network 78.85 MB: 1157.2963 ms (std: 46.0382 ms, min: 1125.9507 ms)
Time taken to load network 199.82 MB: 3076.4666 ms (std: 168.6508 ms, min: 2908.4124 ms)
```
## Performance Analysis
The contribution of each of the three modifications to the performance change can be estimated:
1. Cached Property: (3231.93 -> 2746.35, delta: 485.58) => 485.58 / (485.58+492.43+155.47) = **43%** 
2. Preliminary check: (3231.93 -> 2739.50, delta: 492.43) => 492.43 / (485.58+492.43+155.47) = **43%**
3. map() function: (3231.93 -> 3076.46, delta: 155.47) => 155.47 / (485.58+492.43+155.47) = **14%**
4. 
## Conclusion
Each of these three modifications provided effective performance improvements.
Together, they contributed to the previously mentioned 35% reduction in load time.